Collision detection attempt
Yeah i'm lost, trying to make it so whichever direction pacman is heading, it'll check 10 pixels ahead if that element is a wall (just for now). Want to make it so it compares the colour on the page but have no idea.

Saw something like this: https://stackoverflow.com/questions/2377696/how-can-i-compare-two-color-values-in-jquery-javascript